Adiciona estrutura da classe Pedido
