+++ TecnicasRefatoracao/R01.ExtractMethod/antes/Relatorio.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
+using TecnicasRefatoracao.R01.ExtractMethod.antes;
 
 namespace refatoracao.Aula01.R01.ExtractMethod.antes
 {
@@ -33,28 +34,6 @@ namespace refatoracao.Aula01.R01.ExtractMethod.antes
             Console.WriteLine("*****************************");
             Console.WriteLine("nome: " + pedido.Cliente);
             Console.WriteLine("valor: " + total);
-        }
-    }
-
-    class Pedido
-    {
-        private readonly IList<Item> itens = new List<Item>();
-        public IReadOnlyCollection<Item> Itens
-        {
-            get { return new ReadOnlyCollection<Item>(itens); }
-        }
-
-        private readonly string cliente;
-        public string Cliente { get => cliente; }
-
-        public Pedido(string cliente)
-        {
-            this.cliente = cliente;
-        }
-
-        public void AddItem(string descricao, int quantidade, decimal precoBase, decimal acrescimo, decimal desconto)
-        {
-            itens.Add(new Item(descricao, quantidade, precoBase, acrescimo, desconto));
         }
     }
